refactor: Remove Navigation and Root components
The Navigation and Root components were deleted as they are no longer needed in the application. These components were replaced with UserDetails, EditEventButton, EventDetails, and EventDeleteButton components, which provide improved functionality for user and event details. This change also involved updating the routing in App.jsx and main.jsx files.