refactor: Update App.jsx to use ChakraProvider and extendTheme
The App.jsx file was modified to import ChakraProvider and extendTheme from the @chakra-ui/react package. The ChakraProvider component was added as a wrapper around the existing JSX code, and a new theme object was created using extendTheme to customize the global styles. This change improves the styling and theming of the application.